feat: raise the shared token budget to 10k and share it with simulation
`DEFAULT_TARGET_MAX_TOKENS` was 5000, set when the default model was not a
reasoning model. `max_completion_tokens` covers hidden reasoning as well as
the visible answer, so the same figure now leaves materially less room for
output — truncation shows up as a judge PARSE error carrying billed usage
and no verdict, or as `finish_reason=length` before a tool call.

Simulation had its own `DEFAULT_MAX_TOKENS = 8192` literal; it now falls
through to the shared constant, keeping `EVALUATORQ_LLM_MAX_TOKENS` as the
override. `create_model_job` had a third literal 5000 and now reads the
constant too.

tests/unit/test_shared_defaults.py (renamed from test_default_model_centralized)
pins the budget alongside the model, so the two cannot split again.
