+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to `evaluatorq` are documented here.
 ### Notable defaults
 
 - **One default model for every surface: `openai/gpt-5.6-luna`.** `evaluatorq.contracts.DEFAULT_PIPELINE_MODEL` is now the single source; `simulation.types.DEFAULT_MODEL`, `llm_jury.DEFAULT_JUDGE_MODEL` and the dashboard's `DEFAULT_APPLY_MODEL` are aliases of it rather than three independently drifting literals (previously `gpt-5-mini`, `openai/gpt-5.4-mini`, `openai/gpt-5.4-mini` and `gpt-5.6-luna`). Each surface stays individually overridable — `--attack-model` / `--evaluator-model`, `--sim-model`, `judges=` / `model=`, `EVALUATORQ_APPLY_MODEL` — only the fallback is shared. Every default is now provider-prefixed for the Orq router; targeting OpenAI directly (only `OPENAI_API_KEY` set) means passing the bare `gpt-5.6-luna`, which the red-team defaults did not previously require.
+- **`DEFAULT_TARGET_MAX_TOKENS` is `10_000`, up from `5000`, and simulation now shares it.** On a reasoning model the budget covers hidden reasoning as well as the visible answer, so the old figure left much less room for output than it used to — truncation surfaces as a judge `PARSE` error (`incomplete_details.reason=max_output_tokens`) or, in simulation, `finish_reason=length` before any tool call. It caps `LLMConfig.max_tokens` (attacker), `EvaluatorConfig.max_tokens` (judges), `OpenAIModelTarget`, `create_model_job`, and now simulation's `DEFAULT_MAX_TOKENS`, whose `EVALUATORQ_LLM_MAX_TOKENS` override is unchanged (its default was a separate `8192`).
 - The dashboard's apply-recommendations merge no longer has its own default model. It was deliberately set apart from the pipeline's — rewriting production agent instructions warrants a stronger model than scoring does — and now shares the fallback like every other surface. `EVALUATORQ_APPLY_MODEL` is how you raise it again, unchanged.
 - `red_team()` takes `recommendations=` instead of `generate_recommendations=`, and simulation gained the same flag with the same three forms: `True` (defaults), `False` (skip the LLM call), or a config instance — `RedTeamRecommendationConfig` / `SimulationRecommendationConfig`, both bounded and `extra='forbid'`. **Two removals with no deprecation shim:** `generate_recommendations=` on `red_team()`, and the long-deprecated `config=` alias for `llm_config=` (the shim had targeted removal in 1.4.0 and outlived it). `generate_focus_area_recommendations()` likewise takes `recommendations=` in place of its `max_areas` / `max_traces` pair. Rename the keyword at the call site; behaviour is unchanged. (RES-1286)
 - Simulation now generates remediation suggestions **in-run** rather than from a CLI post-run hook, so a saved run carries them regardless of caller. `simulate()` / `generate_and_simulate()` default the flag to `False`, because the returned `SimulationResult` list has nowhere to carry suggestions — with `save`/`report` both unset the run would pay for them and drop them, which now logs a warning. `eq sim run` and `eq redteam run` keep `--recommendations` on by default. (RES-1286)
